style(all): move  to strategies folder
